search-time/requirements.txt
numpy
faiss-gpu
tqdm
torch